+++ hooks/post_gen_project.py
@@ -328,6 +328,12 @@ def main():
         if "{{ cookiecutter.use_docker }}".lower() == "y":
             remove_node_dockerfile()
 
+    if "{{ cookiecutter.cloud_provider}}".lower() == "none":
+        print(
+            WARNING + "You chose not to use a cloud provider, "
+            "media files won't be served in production." + TERMINATOR
+        )
+
     if "{{ cookiecutter.use_celery }}".lower() == "n":
         remove_celery_files()
         if "{{ cookiecutter.use_docker }}".lower() == "y":