+++ tests/test_cookiecutter_generation.py
@@ -7,11 +7,11 @@ import sh
 import yaml
 from binaryornot.check import is_binary
 
-PATTERN = "{{(\s?cookiecutter)[.](.*?)}}"
+PATTERN = r"{{(\s?cookiecutter)[.](.*?)}}"
 RE_OBJ = re.compile(PATTERN)
 
 YN_CHOICES = ["y", "n"]
-CLOUD_CHOICES = ["AWS", "GCE"]
+CLOUD_CHOICES = ["AWS", "GCE", "None"]
 
 
 @pytest.fixture